@@ -1,15 +1,16 @@
 module github.com/google/trillian-examples
 
-go 1.17
+go 1.19
 
 require (
-	github.com/apache/beam/sdks/v2 v2.0.0-20211013024802-71d7213d9867
-	github.com/cenkalti/backoff/v4 v4.1.3
+	cloud.google.com/go/storage v1.27.0 // indirect
+	github.com/apache/beam/sdks/v2 v2.43.0
+	github.com/cenkalti/backoff/v4 v4.2.0
 	github.com/dsoprea/go-ext4 v0.0.0-20190528173430-c13b09fc0ff8
 	github.com/go-sql-driver/mysql v1.6.0
 	github.com/golang/glog v1.0.0
 	github.com/golang/mock v1.6.0
-	github.com/google/certificate-transparency-go v1.1.2-0.20210512142713-bed466244fa6
+	github.com/google/certificate-transparency-go v1.1.4
 	github.com/google/go-cmp v0.5.9
 	github.com/google/go-github/v39 v39.2.0
 	github.com/google/trillian v1.5.1-0.20220920003623-a57fed5b3a2c
