@@ -1,6 +1,6 @@
 module github.com/google/trillian-examples/serverless/deploy/github/distributor/combine_witness_signatures
 
-go 1.17
+go 1.19
 
 require (
 	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b
