@@ -1,6 +1,6 @@
 module github.com/gcp_serverless_example_module
 
-go 1.17
+go 1.19
 
 require cloud.google.com/go/storage v1.21.0
 
