@@ -15,7 +15,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"strings"
 
 	"github.com/dsoprea/go-ext4"
@@ -173,5 +172,5 @@ func (d *Partition) ReadAll(fullPath string) (buf []byte, err error) {
 	en := ext4.NewExtentNavigatorWithReadSeeker(d, inode)
 	r := ext4.NewInodeReader(en)
 
-	return ioutil.ReadAll(r)
+	return io.ReadAll(r)
 }
